Add test for simultaneous contested election
- Add method to tester to start all servers and provoke a simultaneous
"contested" election.  Implement the contested part by starting all
servers, crashing one server, then sending ElecTimeout to two of the
remaining servers.  If the timing is close enough, a split-vote election
will be observed in which each of the two candidates gains two votes but
is unable to gain the third swing vote from the crashed server.